Decompose rust_admin pipeline.rs into stage submodules
pipeline/ now holds process (transcode stage), final_quote (cost
aggregation for approval), and publish (Autonomi upload + original
file handling). Pure move with module-root glob re-exports.
